// src/models/UsuarioLogin.ts
//interface UsuarioLogin {
    //[x: string]: string;
    //id: number;
    //nome: string
    //usuario: string
    //senha: string;
    //foto: string
    //token: string | null
  //}
  //
  //export default UsuarioLogin


  interface UsuarioLogin {
    id: number;
    nome: string
    usuario: string
    senha: string;
    foto: string
    token: string
  }
  
  export default UsuarioLogin